Create palstrings.cpp
A string is called a Palpal string if it can be divided into contiguous substrings such that:

Each character of the whole string belongs to exactly one substring.
Each of these substrings is a palindrome with length greater than 1.
For example, "abbbaddzcz" is a Palpal string, since it can be divided into "abbba", "dd" and "zcz".

You are given a string s. You may rearrange the characters of this string in any way you choose. Find out if it is possible to rearrange them in such a way that you obtain a Palpal string